fixed backend routes and tested with endpoints

+++ BackEnd/services/customer.service.js
@@ -8,7 +8,7 @@ async function checkIfCustomerExists(email) {
     const query = "SELECT * FROM customer_identifier WHERE customer_email = ?";
     const [rows] = await connection.query(query, [email]);
 
-    return rows.length > 0;
+    return rows;
   } catch (error) {
     console.error("Error checking customer existence:", error);
     throw new Error(
@@ -134,6 +134,13 @@ async function updateCustomer(customer) {
   try {
     const customer_id = customer.customer_id;
 
+    // Replace undefined values with null
+    const customer_email = customer.customer_email || null;
+    const customer_phone_number = customer.customer_phone_number || null;
+    const customer_first_name = customer.customer_first_name || null;
+    const customer_last_name = customer.customer_last_name || null;
+    const active_customer_status = customer.active_customer_status || null;
+
     const query1 = `
       UPDATE customer_identifier
       SET customer_email = ?, customer_phone_number = ?
@@ -147,15 +154,15 @@ async function updateCustomer(customer) {
     `;
 
     const result1 = await connection.query(query1, [
-      customer.customer_email,
-      customer.customer_phone_number,
+      customer_email,
+      customer_phone_number,
       customer_id,
     ]);
 
     const result2 = await connection.query(query2, [
-      customer.customer_first_name,
-      customer.customer_last_name,
-      customer.active_customer_status,
+      customer_last_name,
+      customer_first_name,
+      active_customer_status,
       customer_id,
     ]);
 
@@ -169,6 +176,10 @@ async function updateCustomer(customer) {
 // Delete customer by ID
 async function deleteCustomer(customer_id) {
   try {
+    if (!customer_id) {
+      throw new Error('Customer ID is undefined');
+    }
+    
     const query1 = "DELETE FROM customer_info WHERE customer_id = ?";
     const query2 = "DELETE FROM customer_identifier WHERE customer_id = ?";
